+++ test/api/dataPasses.test.js
@@ -15,12 +15,11 @@ const { expect } = require('chai');
 const request = require('supertest');
 const { server } = require('../../lib/application');
 const { resetDatabaseContent } = require('../utilities/resetDatabaseContent.js');
-const { DetectorType } = require('../../lib/domain/enums/DetectorTypes');
-const { BkpRoles } = require('../../lib/domain/enums/BkpRoles');
 
 const LHC22b_apass1 = {
     id: 1,
     name: 'LHC22b_apass1',
+    skimmingStage: null,
     versions: [
         {
             id: 1,
@@ -41,6 +40,7 @@ const LHC22b_apass1 = {
 const LHC22b_apass2 = {
     id: 2,
     name: 'LHC22b_apass2',
+    skimmingStage: null,
     versions: [
         {
             id: 2,
@@ -58,31 +58,11 @@ const LHC22b_apass2 = {
     simulationPassesCount: 1,
 };
 
-const LHC22a_apass1 = {
-    id: 3,
-    name: 'LHC22a_apass1',
-    versions: [
-        {
-            id: 3,
-            dataPassId: 3,
-            description: 'Some random desc for apass 1',
-            reconstructedEventsCount: 50848111,
-            outputSize: 55761110122610,
-            lastSeen: 105,
-            deletedFromMonAlisa: false,
-            createdAt: 1704884400000,
-            updatedAt: 1704884400000,
-        },
-    ],
-    runsCount: 4,
-    simulationPassesCount: 2,
-};
-
 module.exports = () => {
     before(resetDatabaseContent);
 
     describe('GET /api/dataPasses', () => {
-        it('should successfuly fetch all data', (done) => {
+        it('should successfully fetch all data', (done) => {
             request(server)
                 .get('/api/dataPasses')
                 .expect(200)
@@ -93,14 +73,14 @@ module.exports = () => {
                     }
 
                     const { data, meta } = res.body;
-                    expect(meta).to.be.eql({ page: { totalCount: 3, pageCount: 1 } });
+                    expect(meta).to.be.eql({ page: { totalCount: 5, pageCount: 1 } });
                     expect(data).to.be.an('array');
-                    expect(data).to.be.lengthOf(3);
+                    expect(data).to.be.lengthOf(5);
 
                     done();
                 });
         });
-        it('should successfuly filter on ids', (done) => {
+        it('should successfully filter on ids', (done) => {
             request(server)
                 .get('/api/dataPasses?filter[ids][]=1')
                 .expect(200)
@@ -119,7 +99,7 @@ module.exports = () => {
                     done();
                 });
         });
-        it('should successfuly filter on names', (done) => {
+        it('should successfully filter on names', (done) => {
             request(server)
                 .get('/api/dataPasses?filter[names][]=LHC22b_apass2')
                 .expect(200)
@@ -137,7 +117,7 @@ module.exports = () => {
                     done();
                 });
         });
-        it('should retrive no records when filtering on ids', (done) => {
+        it('should retrieve no records when filtering on ids', (done) => {
             request(server)
                 .get('/api/dataPasses?filter[ids][]=9999')
                 .expect(200)
@@ -154,7 +134,7 @@ module.exports = () => {
                     done();
                 });
         });
-        it('should retrive no records when filtering on names', (done) => {
+        it('should retrieve no records when filtering on names', (done) => {
             request(server)
                 .get('/api/dataPasses?filter[names][]=LHC22b_aasdfpass2asdf')
                 .expect(200)
@@ -170,7 +150,7 @@ module.exports = () => {
                     done();
                 });
         });
-        it('should succefully filter on ids given as array', (done) => {
+        it('should successfully filter on ids given as array', (done) => {
             request(server)
                 .get('/api/dataPasses?filter[ids][]=1&filter[ids][]=2')
                 .expect(200)
@@ -186,7 +166,7 @@ module.exports = () => {
                     done();
                 });
         });
-        it('should succefully filter on lhcPeriodIds', (done) => {
+        it('should successfully filter on lhcPeriodIds', (done) => {
             request(server)
                 .get('/api/dataPasses?filter[lhcPeriodIds][]=2')
                 .expect(200)
@@ -203,7 +183,7 @@ module.exports = () => {
                     done();
                 });
         });
-        it('should succefully filter on simulationPassIds', (done) => {
+        it('should successfully filter on simulationPassIds', (done) => {
             request(server)
                 .get('/api/dataPasses?filter[simulationPassIds][]=1')
                 .expect(200)
@@ -219,7 +199,7 @@ module.exports = () => {
                     done();
                 });
         });
-        it('should successfuly sort on id and name', (done) => {
+        it('should successfully sort on id and name', (done) => {
             request(server)
                 .get('/api/dataPasses?sort[id]=DESC&sort[name]=ASC')
                 .expect(200)
@@ -231,11 +211,13 @@ module.exports = () => {
 
                     const { data: dataPasses } = res.body;
                     expect(dataPasses).to.be.an('array');
-                    expect(dataPasses).to.be.lengthOf(3);
-                    expect(dataPasses).to.have.ordered.deep.members([
-                        LHC22a_apass1,
-                        LHC22b_apass2,
-                        LHC22b_apass1,
+                    expect(dataPasses).to.be.lengthOf(5);
+                    expect(dataPasses.map(({ name }) => name)).to.have.ordered.members([
+                        'LHC22a_apass2_skimmed',
+                        'LHC22a_skimming',
+                        'LHC22a_apass1',
+                        'LHC22b_apass2',
+                        'LHC22b_apass1',
                     ]);
 
                     done();
@@ -253,15 +235,17 @@ module.exports = () => {
 
                     const { data: dataPasses } = res.body;
                     expect(dataPasses).to.be.an('array');
-                    expect(dataPasses).to.have.ordered.deep.members([
-                        LHC22b_apass2,
-                        LHC22b_apass1,
+                    expect(dataPasses.map(({ name }) => name)).to.have.ordered.deep.members([
+                        'LHC22a_skimming',
+                        'LHC22a_apass1',
+                        'LHC22b_apass2',
+                        'LHC22b_apass1',
                     ]);
 
                     done();
                 });
         });
-        it('should return 400 when bad query paramter provided', (done) => {
+        it('should return 400 when bad query parameter provided', (done) => {
             request(server)
                 .get('/api/dataPasses?a=1')
                 .expect(400)
@@ -308,51 +292,6 @@ module.exports = () => {
                     expect(titleError.detail).to.equal('"query.page.limit" must be greater than or equal to 1');
                     done();
                 });
-        });
-    });
-
-    describe('POST /api/dataPasses/gaqDetectors', () => {
-        it('should succesfuly set GAQ detectors', async () => {
-            const dataPassId = 3;
-            const runNumbers = [49, 56];
-            const detectorIds = [4, 7];
-            const response = await request(server).post(`/api/dataPasses/gaqDetectors?token=${BkpRoles.GAQ}`).send({
-                dataPassId,
-                runNumbers,
-                dplDetectorIds: detectorIds,
-            });
-            expect(response.status).to.be.equal(201);
-            expect(response.body.data).to.have.all.deep.members(runNumbers
-                .flatMap((runNumber) => detectorIds.map((detectorId) => ({ dataPassId, runNumber, detectorId }))));
-        });
-
-        it('should fail to set GAQ detectors because of insufficient permission', async () => {
-            const dataPassId = 3;
-            const runNumbers = [49, 56];
-            const detectorIds = [4, 7];
-            const response = await request(server).post(`/api/dataPasses/gaqDetectors?token=${BkpRoles.GUEST}`).send({
-                dataPassId,
-                runNumbers,
-                dplDetectorIds: detectorIds,
-            });
-            expect(response.status).to.be.equal(403);
-
-            const { errors } = response.body;
-            expect(errors.find(({ title }) => title === 'Access denied')).to.not.be.null;
-        });
-    });
-
-    describe('GET /api/dataPasses/gaqDetectors', () => {
-        it('should return 200 with the list of GAQ detectors', async () => {
-            const response = await request(server).get('/api/dataPasses/gaqDetectors?dataPassId=3&runNumber=56');
-
-            expect(response.status).to.equal(200);
-            const { data } = response.body;
-            expect(data).to.be.an('array');
-            expect(data).to.have.all.deep.members([
-                { id: 7, name: 'FT0', type: DetectorType.PHYSICAL },
-                { id: 4, name: 'ITS', type: DetectorType.PHYSICAL },
-            ]);
         });
     });
 };