+++ test/lib/server/services/dataPasses/DataPassService.test.js
@@ -16,12 +16,11 @@ const { resetDatabaseContent } = require('../../../../utilities/resetDatabaseCon
 const assert = require('assert');
 const { NotFoundError } = require('../../../../../lib/server/errors/NotFoundError.js');
 const { dataPassService } = require('../../../../../lib/server/services/dataPasses/DataPassService.js');
-const { BadParameterError } = require('../../../../../lib/server/errors/BadParameterError.js');
-const { DetectorType } = require('../../../../../lib/domain/enums/DetectorTypes.js');
 
 const LHC22b_apass1 = {
     id: 1,
     name: 'LHC22b_apass1',
+    skimmingStage: null,
     versions: [
         {
             id: 1,
@@ -42,6 +41,7 @@ const LHC22b_apass1 = {
 const LHC22b_apass2 = {
     id: 2,
     name: 'LHC22b_apass2',
+    skimmingStage: null,
     versions: [
         {
             id: 2,
@@ -62,6 +62,7 @@ const LHC22b_apass2 = {
 const LHC22a_apass1 = {
     id: 3,
     name: 'LHC22a_apass1',
+    skimmingStage: null,
     versions: [
         {
             id: 3,
@@ -70,7 +71,7 @@ const LHC22a_apass1 = {
             reconstructedEventsCount: 50848111,
             outputSize: 55761110122610,
             lastSeen: 105,
-            deletedFromMonAlisa: false,
+            deletedFromMonAlisa: true,
             createdAt: 1704884400000,
             updatedAt: 1704884400000,
         },
@@ -83,19 +84,19 @@ module.exports = () => {
     before(resetDatabaseContent);
 
     describe('Fetching', () => {
-        it('should succesfully get by id', async () => {
+        it('should successfully get by id', async () => {
             const dataPass = await dataPassService.getByIdentifier({ id: 1 });
             expect(dataPass).to.be.eql(LHC22b_apass1);
         });
 
-        it('should succesfully get by name', async () => {
+        it('should successfully get by name', async () => {
             const dataPass = await dataPassService.getByIdentifier({ name: 'LHC22a_apass1' });
             expect(dataPass).to.be.eql(LHC22a_apass1);
         });
 
-        it('should succesfully get all data', async () => {
+        it('should successfully get all data', async () => {
             const { rows: dataPasses } = await dataPassService.getAll();
-            expect(dataPasses).to.be.lengthOf(3);
+            expect(dataPasses).to.be.lengthOf(5);
         });
 
         it('should fail when no Data Pass with given id', async () => {
@@ -105,7 +106,7 @@ module.exports = () => {
             );
         });
 
-        it('should succesfully filter data passes on names', async () => {
+        it('should successfully filter data passes on names', async () => {
             const dto = {
                 query: {
                     filter: {
@@ -118,7 +119,7 @@ module.exports = () => {
             expect(dataPasses[0]).to.be.eql(LHC22b_apass1);
         });
 
-        it('should succesfully filter data passes on ids', async () => {
+        it('should successfully filter data passes on ids', async () => {
             const dto = {
                 query: {
                     filter: {
@@ -134,7 +135,7 @@ module.exports = () => {
             expect(await dataPassService.getByIdentifier({ id: 99999 })).to.be.null;
         });
 
-        it('should succesfully filter data passes on lhc petriods ids', async () => {
+        it('should successfully filter data passes on lhc periods ids', async () => {
             const dto = {
                 query: {
                     filter: {
@@ -147,7 +148,7 @@ module.exports = () => {
             expect(dataPasses).to.have.deep.members([LHC22b_apass1, LHC22b_apass2]);
         });
 
-        it('should succesfully filter data passes on simulation pass ids', async () => {
+        it('should successfully filter data passes on simulation pass ids', async () => {
             const dto = {
                 query: {
                     filter: {
@@ -159,7 +160,7 @@ module.exports = () => {
             expect(dataPasses).to.have.all.deep.members([LHC22b_apass1, LHC22b_apass2]);
         });
 
-        it('should succesfully sort data passes by names', async () => {
+        it('should successfully sort data passes by names', async () => {
             const dto = {
                 query: {
                     sort: {
@@ -168,39 +169,8 @@ module.exports = () => {
                 },
             };
             const { rows: dataPasses } = await dataPassService.getAll(dto.query);
-            expect(dataPasses).to.have.ordered.deep.members([LHC22a_apass1, LHC22b_apass1, LHC22b_apass2]);
-        });
-    });
-    describe('Manage GAQ detectors', () => {
-        const dataPassId = 3;
-        it('should successfuly set GAQ detectors', async () => {
-            const runNumbers = [49, 56];
-            const detectorIds = [4, 7];
-            const data = await dataPassService.setGaqDetectors(dataPassId, runNumbers, detectorIds);
-            expect(data).to.be.have.all.deep.members(runNumbers
-                .flatMap((runNumber) => detectorIds.map((detectorId) => ({ dataPassId, runNumber, detectorId }))));
-        });
-        it('should fail to set GAQ detectors because of miaaing association', async () => {
-            let errorMessage = `No association between data pass with id ${dataPassId} and following runs: 1`;
-            assert.rejects(
-                () => dataPassService.setGaqDetectors(dataPassId, [1], [4]),
-                new BadParameterError(errorMessage),
-            );
-            errorMessage = `No association between runs and detectors: ${JSON.stringify([[56, 'CPV']])}`;
-            assert.rejects(
-                () => dataPassService.setGaqDetectors(dataPassId, [105, 56], [1]),
-                new BadParameterError(errorMessage),
-            );
-        });
-
-        it('should get GAQ detectors', async () => {
-            const detectors = await dataPassService.getGaqDetectors(3, 56);
-            expect(detectors).to.be.an('array');
-            expect(detectors).to.be.lengthOf(2);
-            expect(detectors).to.have.all.deep.members([
-                { id: 7, name: 'FT0', type: DetectorType.PHYSICAL },
-                { id: 4, name: 'ITS', type: DetectorType.PHYSICAL },
-            ]);
+            expect(dataPasses.map(({ name }) => name)).to.have
+                .ordered.members(['LHC22a_apass1', 'LHC22a_apass2_skimmed', 'LHC22a_skimming', 'LHC22b_apass1', 'LHC22b_apass2']);
         });
     });
 };