+++ test/lib/server/services/statistics/StatisticsService.test.js
@@ -337,7 +337,7 @@ module.exports = () => {
         ]);
     });
 
-    it('should successfully filter out environment history before a date (included) for history occurences', async () => {
+    it('should successfully filter out environment history before a date (included) for history occurrences', async () => {
         {
             const historyOccurrences = await statisticsService.getHistoryOccurrencesInEnvironments({
                 from: new Date('2019/08/09 14:30:00').getTime(),
@@ -355,7 +355,7 @@ module.exports = () => {
         }
     });
 
-    it('should successfully filter out environment history after a date (excluded) for history occurences', async () => {
+    it('should successfully filter out environment history after a date (excluded) for history occurrences', async () => {
         {
             const historyOccurrences = await statisticsService.getHistoryOccurrencesInEnvironments({
                 from: new Date('2019/08/09 14:15:00').getTime(),